pyproject.toml
[tool.poetry]
name = "authpi"
package-mode = false
version = "0.0.1"
description = "开放的身份云API实现，参考了Authing"
authors = ["XYCode Kerman <xycode-xyc@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = {extras = ["full"], version = "^0.111.0"}
uvicorn = "^0.29.0"
odmantic = "^1.0.2"
motor = "^3.4.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.230.0"}
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.21.2"
httpx = "^0.27.0"
asgi-lifespan = "^2.1.0"
faker = "^25.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
